Fix parameter metadata bug. Move DSISetAction to action package.
