+++ build.gradle
@@ -5,7 +5,7 @@ subprojects {
     apply plugin: 'maven'
 
     group 'org.iot-dsa'
-    version '0.49.1'
+    version '0.49.2'
 
     sourceCompatibility = 1.8
     targetCompatibility = 1.8
